+++ src/auth/authSlice.ts
@@ -15,16 +15,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { anonAuth } from "./authConstants";
 import { AuthState } from "./authTypes";
+import { localStorage } from "redux-persist-webextension-storage";
 
-export const anonAuth: AuthState = Object.freeze({
-  userId: undefined,
-  email: undefined,
-  isLoggedIn: false,
-  isOnboarded: false,
-  extension: true,
-  scope: null,
-  flags: [],
-  organizations: [],
-  groups: [],
+export const authSlice = createSlice({
+  name: "auth",
+  initialState: anonAuth,
+  reducers: {
+    setAuth: (state, { payload }: PayloadAction<AuthState>) => payload,
+  },
 });
+
+export const persistAuthConfig = {
+  key: "authOptions",
+  storage: localStorage,
+};
+
+export const authActions = authSlice.actions;
