+++ src/core.ts
@@ -830,44 +830,6 @@ export interface IconConfig {
 
 export type UserOptions = Record<string, Primitive>;
 
-export type OrganizationAuthState = {
-  readonly id: string;
-  // XXX: why is this optional?
-  readonly name?: string;
-  readonly scope?: string;
-};
-
-export type AuthState = {
-  readonly userId?: string | null;
-
-  readonly email?: string | null;
-
-  readonly scope?: string | null;
-
-  readonly isLoggedIn: boolean;
-
-  readonly isOnboarded: boolean;
-
-  readonly extension: boolean;
-
-  readonly organization?: OrganizationAuthState | null;
-
-  readonly organizations: Array<{
-    id: UUID;
-    name: string;
-  }>;
-
-  readonly groups: Array<{
-    id: UUID;
-    name: string;
-  }>;
-
-  /**
-   * List of feature flags for the user.
-   */
-  readonly flags: string[];
-};
-
 export type RawConfig = {
   kind: "service" | "extensionPoint" | "component" | "reader" | "recipe";
   metadata: Metadata;
